Pathfinding fix + drive-by improvements
A lot of the changes in this commit are from drive-by debugging
improvements while debugging a pathfinding failure on game #14652.

The root cause was that, when we find a goal, our heuristic changes, so
we need to update our frontier priorities and, importantly, *re-insert*
the current goal node to re-evaluate it at the right priority with the
updated f-scores.

Other drive-by improvements/fixes:
- priority queue using Reverse, so no longer need to insert negative f
  scores
- ant logic was working off of precomputed 'tick_offset', which
  only being updated right before computing the cost in visit, so wasn't
  up-to-date after visit is called
- added pathfinding verify_path logic, checks for pathfinding bugs
  always now, instead of only encountering them on edges we visit during
  simulation
- verify solution right after generation in solver
- fix server eval mode, should only pass hyperparams file when sweeping
- show per-tick updates using Pos, for consistency of x/y vs. row/col
  ordering
- commented-out verification logic in graph wasn't up-to-date, fixed
- debug printing for pathfinding States